moved test functionality to new file
